# static/swagger.json
{"info": {"description": "This is ASAR swagger api documentation", "version": "1.0.0", "contact": {"email": "aboelkassem.me@gmail.com"}, "license": {"name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0.html"}, "title": "ASAR API"}, "servers": [{"description": "ASAR server", "url": "http://127.0.0.1:5000"}], "paths": {"/api/auth/login": {"post": {"parameters": [{"in": "body", "required": false, "name": "login", "schema": {"$ref": "#/components/schemas/Login"}, "description": "Login DTO model"}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}}}}, "/api/auth/register": {"post": {"parameters": [{"in": "body", "required": false, "name": "register", "schema": {"$ref": "#/components/schemas/Register"}, "description": "Register DTO model"}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}}}}, "/api/auth/logout": {"post": {"responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}}}}, "/api/predict": {"post": {"parameters": [{"in": "body", "required": false, "name": "image", "schema": {"$ref": "#/components/schemas/Predict"}, "description": "image format should be base64 image, for example in body {'image' 'data:image/png;base64,.....'}"}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}}}}, "/api/delete_result/{prediction_id}": {"post": {"parameters": [{"in": "path", "name": "prediction_id", "type": "int", "description": "the prediction_id to be deleted", "required": true}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}}}}, "/api/prediction/{prediction_id}": {"get": {"parameters": [{"in": "path", "name": "prediction_id", "type": "int", "description": "the prediction_id to get details", "required": true}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PredictionDetails"}}}}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Basic"}}}}}}}, "/api/home": {"get": {"responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PredictionList"}}}}}}}}, "openapi": "3.0.2", "components": {"schemas": {"Login": {"type": "object", "properties": {"username": {"type": "string"}, "password": {"type": "string"}}}, "Basic": {"type": "object", "properties": {"message": {"type": "string"}}}, "Register": {"type": "object", "properties": {"confirm_password": {"type": "string"}, "username": {"type": "string"}, "password": {"type": "string"}}}, "Predict": {"type": "object", "properties": {"image": {"type": "string"}}}, "PredictionDetails": {"type": "object", "properties": {"id": {"type": "integer"}, "result": {"type": "string"}, "image": {"type": "string"}}}, "PredictionList": {"type": "object", "properties": {"results": {"type": "array", "items": {"$ref": "#/components/schemas/PredictionDetails"}}}}}}}